+++ backend/src/external-connectors/gateway.js
@@ -17,6 +17,7 @@ import { GitHubCommunityConnector } from './connectors/github-community-connecto
 import { CWGCConnector } from './connectors/cwgc-connector.js';
 import { FindAGraveConnector } from './connectors/findagrave-connector.js';
 import { DeceasedOnlineConnector } from './connectors/deceased-online-connector.js';
+import { KuburSGConnector } from './connectors/kubur-sg-connector.js';
 import { evaluateLicense } from './licensing.js';
 import { handleFailure, createFallbackResponse } from './failure-handler.js';
 import { writeAuditEntry, createSuccessAudit, createFailureAudit } from './audit-log.js';
@@ -53,6 +54,9 @@ function getConnector(sourceId) {
       break;
     case 'uk-deceased-online':
       connector = new DeceasedOnlineConnector();
+      break;
+    case 'kubur-sg':
+      connector = new KuburSGConnector();
       break;
     default:
       return null;